Move non-clean functions out of clean
